requirements.txt
amqp==1.4.9
anyjson==0.3.3
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.3.0.23
celery==3.1.18
certifi==2018.11.29
chardet==3.0.4
coverage==4.0.3
Django==2.1.5
django-paypal==0.5.0
djangorestframework==3.9.1
idna==2.8
kombu==3.0.37
model-mommy==1.6.0
more-itertools==5.0.0
Pillow==5.4.1
pluggy==0.8.1
py==1.7.0
python-coveralls==2.9.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
